Add condition to only attempt question deletion if answers are deleted first

+++ src/main/java/com/codecool/stackoverflowtw/service/QuestionService.java
@@ -53,6 +53,10 @@ public class QuestionService {
     }
 
     public boolean deleteQuestionById(int id) {
+        boolean answersDeleted = answersDao.deleteAnswersByQuestionId(id);
+        if (!answersDeleted) {
+            return false;
+        }
         return questionsDAO.deleteQuestionByQuestionId(id);
     }
